fixup: Merge Builder extension method classes
* Prefer AgentsAppBuilderExtensions as name, due to matching the incoming `this` type.
